+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 1. [#283](https://github.com/influxdata/influxdb-client-js/pull/283): Allow to specify custom HTTP headers in WriteApi or QueryApi.
 1. [#287](https://github.com/influxdata/influxdb-client-js/pull/287): Throw HttpError with code and message from response body.
 1. [#289](https://github.com/influxdata/influxdb-client-js/pull/289): Support `deno`.
+1. [#292](https://github.com/influxdata/influxdb-client-js/pull/289): Add query.deno.ts example.
 
 ### Breaking Changes